Added links to the http://spelunky.wikia.com/wiki/Spelunky_Wiki

+++ source/sprites/sprite_utils.cpp
@@ -8,9 +8,7 @@
 
 namespace sprite_utils {
 
-    void set_entry_xy(SpriteInfo *i, int x, int y) {
-//        if(i == nullptr)
-//            printf("NULLPTR XY\n");
+    void set_entry_xy(SpriteInfo *i, u16 x, u16 y) {
         i->entry->x = x;
         i->entry->y = y;
     }
